requirements.txt
kivymd==1.2.0
plyer==2.1.0
python-dateutil==2.9.0.post0
